checkpoint

+++ app/src/main/java/com/example/desafiogds/Marca.kt
@@ -1,6 +1,6 @@
 package com.example.desafiogds
 
-class Marca( var id : Int, var marca: String) {
+class Marca(var id: Long, var marca: String) {
 
 
 }
